Add the model of the application with ViewModel respond for them
Add Addiction model which will represent actual elements contained in the application.  To every view were created a ViewModel, not everyone has implementations but only empty structure.
Edits the looks of the AddictionVellView to more preferable to develop. Still contains static strings. Has bug that the font is cute in shape of Circle(). Has to recognize how to fit font to the shape size.
Adds NavigationView to help make segues in application. For now only has the name of the app and the button which will be respond for sorting elements in the array. Plus button was changed to Image(systemName:) to fit to sfSymbols.