Add the model and view model of AddictionList
Adds the reference in sceneDelegate to create an ViewModel of main application.
Adds the AddictionListViewModel which contains the basics addiction array.
Deleted the tests which were created by default.
Edit README.md: creating what technologies it's will contain and which language. And make description about how to run an application