Adds view models to views
In this commit main view AddictionListView got his ViewModel as a StateObject to save the status of the view after change of them after make a segue.
Protocol AddictionType conforms to Identifiable to be used as argument in List() usage in AddictionListView. Conforms to protocol allowed usage table of addictions to be used to display individual rows.
Created a func createExampleAddiction() -> Void in AddictionListViewModel to return an example array of addiction, before created a service which allows to withdraw data from Core Data.

Every view has been adopted to his viewModel.
List of bugs in AddictionCellView :
* Duration of rehab is showing as Date.description
* There are text of "HelloWorld" which nothing describe
List of missing data in models:
* Missing description information - it contains short notes of why user drop some addiction.